# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthResult": {
        "description": "Returned by register/login so mobile clients get the session token in JSON\n(the cookie is also set, for the web).",
        "properties": {
          "session_token": {
            "title": "Session Token",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/PublicUser"
          }
        },
        "required": [
          "user",
          "session_token"
        ],
        "title": "AuthResult",
        "type": "object"
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Identity, wallet and AI content generation for TOONTOON.",
    "title": "ARTEKI API",
    "version": "generated"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/forgot-password": {
      "post": {
        "description": "Start a password reset. ALWAYS returns 200 (never reveals whether the\nemail exists). A single-use reset token (TTL 1h) is stored in Redis.\n\nEmail delivery is not wired yet, so — exactly like the magic-link flow — the\ntoken is returned in the JSON response for now. TODO(email): once an ESP is\nconfigured, send the token by email and STOP returning `devToken`/`devLink`.",
        "operationId": "forgot_password_api_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forgot Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/google/callback": {
      "get": {
        "operationId": "google_callback_api_auth_google_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Callback",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/google/login": {
      "get": {
        "description": "Kick off Google OAuth.\n\n``platform=app`` is used by the native app (opens this URL inside\n``ASWebAuthenticationSession``): on success the callback redirects to a\ndeep link (``{app_deep_link_scheme}://auth/callback?token=...``) instead\nof the web frontend, so the app can capture the session token straight\nfrom the URL — no cookie jar needed. Default ``platform=web`` behaves\nexactly like the Boostyfi flow (cookie + redirect to the frontend).",
        "operationId": "google_login_api_auth_google_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "default": "web",
              "title": "Platform",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/login": {
      "post": {
        "description": "Log in with email + password. Returns the session token (JSON) + cookie.",
        "operationId": "login_api_auth_login_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "arteki-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arteki-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/magic-link": {
      "post": {
        "description": "Issue a magic-link token.\n\nEmail is NOT actually sent — the dev link is returned directly in the JSON\nresponse (temporary, until real email delivery is wired up).\n\nПод теми же счётчиками, что и сброс пароля. Без них эта ручка была\nединственным входом без лимита: `verify` заводит аккаунт с бонусом, и в\ndev-режиме, где ссылка приходит в ответе, это был конвейер монет в обход\nлимита на гостей. С почтой это станет бомбёжкой чужого ящика.",
        "operationId": "magic_link_api_auth_magic_link_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Magic Link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/magic-link/verify": {
      "get": {
        "description": "Same magic-link token as ``/verify``, but for native/app clients:\nreturns ``{user, session_token}`` directly in JSON instead of a redirect\nwith ``Set-Cookie`` — no cookie-jar interception needed. The cookie is\nstill set too (harmless if the caller ignores it), so this also works\nfine from a browser if ever needed.",
        "operationId": "magic_link_verify_json_api_auth_magic_link_verify_get",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Magic Link Verify Json",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/oauth/apple": {
      "post": {
        "description": "Verify an Apple ``identity_token`` (from ``ASAuthorizationAppleIDProvider``\non the client) and start a session. Returns ``{user, session_token}`` in\nJSON (same shape as register/login) — no redirect involved, the app talks\nto Apple itself and only hands us the token to verify.",
        "operationId": "apple_auth_api_auth_oauth_apple_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "arteki-session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Arteki-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleAuthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apple Auth",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/reset-password": {
      "post": {
        "description": "Consume a reset token and set a new password (Argon2).",
        "operationId": "reset_password_api_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}